Event fees on Comps (#553)
* WIP

* WIP

* Unique constraint should use parent_id

* Add quantity and available

* WIP commit

* - Removed hold_id from comps routes
- Refactored to reuse code from Holds for Comps
- Removed the cascading delete from hold.parent_hold_id because TicketInstances from child holds must be returned to the pools/holds they came from
- Removed the cascading delete from ticket_instances for when an order_item gets deleted, rather be safe
- Added some documentation to the holds methods
- Removed per item fees from order_items that are attached to a Comp hold
- Removed event fees for orders that only have comps for an event in them
- Moved OrderItem::destroy to Order::destroy so that it check the status of the order before deleting the items
- Added a foreign key ErrorCode in DatabaseError

* Fixed test

* Fixed tests

* Routing and fixing tests

- Renamed PUT to PATCH in /comps/{id} update

* Change Fee items quantity to match the tickets quantity

* Remove `comp_id` from `order_items`

Also:
- Return Quantity when creating a hold

* WIP

* Add backtrace for more debug info.

* Shouldn't have been checked in....

* Apply hold discount

* Fixing paging tests

* Fixed tests

* Fixed Postman

* Neatened up some code

* Neaten up this code

+++ api/src/controllers/ticket_types.rs
@@ -156,7 +156,7 @@ pub fn index(
         )?);
     }
     payload.paging.limit = payload.data.len() as u32;
-    payload.paging.total = payload.data.len() as u32;
+    payload.paging.total = payload.data.len() as u64;
 
     Ok(HttpResponse::Ok().json(&payload))
 }
